+++ changelog.md
@@ -1,3 +1,4 @@
 ### General
 
+- Added rule `legacyProjectileMargin` and `projectileHitThroughReintroduce`
 - Requires carpet 26.1